Function takes input Warg, to test results of differently computed Weight matrices
I had uncertainty if the source space yielded better seperability if we use all columns of the weight matrix. Also we didn't originally project the 'whitening' parameter onto the eigen vector, so I tested the mean accuracy of male-/female groups in three different scenarios:
- No projection; only whitening (W = P' )
- Electrode space (W = identity matrix)
- Whitening and projection (W = P' * V)
The last one is correct one to use in CSP procedure and this showed the best accuracy.
This version is changed to only test for this and not for further development